Add function for switching to a new MAC address
This is useful if you’re on some public wifi hotspot that only allows
you a certain amount of access in a given period. On the cross-country
trains in the UK, this typically means 15 minutes(!) per day for free,
after which you are asked to pay an extortionate amount for continued
(and still flaky) access.

It its current implementation, you may need to remac a few times until a
new MAC address is output. I may try and improve on this by adding a
loop that exits once a new MAC address is generated.